combined Interfaces with classes, added game system
Took all of the interfaces and combined them with the classes that used them. I also removed the folders as they were not needed anymore.

Added a GameSystem file that will handle most of the logic for the game.

Added a settings class for each gametypes settings to go inot (also fo future gametypes)